Move executor.CloneRepository() to a separate file
